Move test files into appropriate directories
